Material to support weighted scoring for all questionnaire types available to an Assignment (Review, Metareview, Author Feedback and Teammate Review).
Weighting is used in the instructor view of the scores.

Students' view of the scores needs to be updated. It currently does not compute the scoring in the same way as the instructor, which will cause problems.

This is a minor project to undertake and will involve rewriting the student_assignment/view_scores action. My recommendation is to use the existing grades/view display but limit it to a student level view (no edit to grade, no access to reviewer names/e-mail addresses)

git-svn-id: http://pg-server.csc.ncsu.edu/svn/pg/trunk@500 392d7b7b-3f31-0410-9dc3-c95a9635ea79