Rewrite of ReviewFeedback. Previous version had lots of extraneous code and did not maintain the same layout in view/edit/new. Several methods and views were no longer used by the system and I've removed them.
Ideally we should be able to create a single Rubric response display for any type (Review/Metareview/Feedback). This is a bigger challenge than it sounds as we have to start from the ground up to get the items in 'sync'. I've tried to tackle this in the past but have never gotten everything complete.

A better approach might be to work on one type at a time until we can get the items in parallel. Once all rubrics use the same basic structure for the display, the next step will be to refactor the code so that they use one single approach.

git-svn-id: http://pg-server.csc.ncsu.edu/svn/pg/trunk@562 392d7b7b-3f31-0410-9dc3-c95a9635ea79